Refactoring for a different approach
Many requests are going to return collections of resources or compound
documents where the resources are are accompanied by included, but
secondary, additional resources. Fundamentally, we need a class method
that can return a collection of objects. We will also need to be be able
to return an object that corresponds in a meaningful but convenient way
to a JSON-API compound document